+++ lib/custom_like_button.dart
@@ -1,5 +1,6 @@
+/// A Like Button with animation is to apply custom(different) icon with color, size and flip animation
 library custom_like_button;
 
-export 'src/custom_like_button.dart';
+export 'src/custom_like_button_with_animation.dart';
 export 'src/animation/animationbuilder.dart';
 export 'src/animation/rotation.dart';