+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/home/addweb/fvm/versions/3.10.3"
+export "FLUTTER_ROOT=/home/addweb/fvm/versions/3.13.7"
 export "FLUTTER_APPLICATION_PATH=/home/addweb/flutterProjects19june/contribution_app/ContributionApp/custom_like_button"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=0.0.1"
-export "FLUTTER_BUILD_NUMBER=0.0.1"
+export "FLUTTER_BUILD_NAME=0.0.3"
+export "FLUTTER_BUILD_NUMBER=0.0.3"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"