+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.3
 
-* TODO: Describe initial release.
+- 0.0.1: initial release of custom like button with animation package..
+- 0.0.2: documentation change.
+- 0.0.3: error resolved for static analysis.